+++ packages/nimble-components/src/table/index.ts
@@ -1,4 +1,4 @@
-import { observable } from '@microsoft/fast-element';
+import { attr, observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
     ColumnDef as TanStackColumnDef,
@@ -9,9 +9,11 @@ import {
     getCoreRowModel as tanStackGetCoreRowModel,
     TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
+import type { TableColumn } from '../table-column/base';
+import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableRecord } from './types';
+import type { TableRecord, TableRowState, TableValidity } from './types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -25,21 +27,29 @@ declare global {
 export class Table<
     TData extends TableRecord = TableRecord
 > extends FoundationElement {
+    @attr({ attribute: 'id-field-name' })
+    public idFieldName?: string | null;
+
     @observable
     public data: TData[] = [];
 
-    // TODO: Temporarily expose the columns as a string array. This will ultimately be
-    // column definitions provided by slotted elements.
+    /**
+     * @internal
+     */
     @observable
-    public columns: string[] = [];
+    public tableData: TableRowState<TData>[] = [];
 
-    // TODO: Temporarily expose the column headers as a string array.
     @observable
-    public columnHeaders: string[] = [];
+    public readonly columns: TableColumn[] = [];
+
+    public get validity(): TableValidity {
+        return this.tableValidator.getValidity();
+    }
 
     private readonly table: TanStackTable<TData>;
     private options: TanStackTableOptionsResolved<TData>;
     private readonly tableInitialized: boolean = false;
+    private readonly tableValidator = new TableValidator();
 
     public constructor() {
         super();
@@ -47,6 +57,13 @@ export class Table<
             data: [],
             onStateChange: (_: TanStackUpdater<TanStackTableState>) => {},
             getCoreRowModel: tanStackGetCoreRowModel(),
+            getRowId: record => {
+                if (this.idFieldName) {
+                    return record[this.idFieldName] as string;
+                }
+                // Return a falsey value to use the default ID from TanStack.
+                return '';
+            },
             columns: [],
             state: {},
             renderFallbackValue: null,
@@ -54,6 +71,15 @@ export class Table<
         };
         this.table = tanStackCreateTable(this.options);
         this.tableInitialized = true;
+    }
+
+    public idFieldNameChanged(
+        _prev: string | undefined,
+        _next: string | undefined
+    ): void {
+        // Force TanStack to detect a data update because a row's ID is only
+        // generated when creating a new row model.
+        this.trySetData([...this.data]);
     }
 
     public dataChanged(
@@ -66,8 +92,32 @@ export class Table<
 
         // Ignore any updates that occur prior to the TanStack table being initialized.
         if (this.tableInitialized) {
-            this.updateTableOptions({ data: this.data });
+            this.trySetData(this.data);
         }
+    }
+
+    public checkValidity(): boolean {
+        return this.tableValidator.isValid();
+    }
+
+    private trySetData(newData: TData[]): void {
+        const areIdsValid = this.tableValidator.validateDataIds(
+            newData,
+            this.idFieldName
+        );
+        if (areIdsValid) {
+            this.updateTableOptions({ data: newData });
+        } else {
+            this.updateTableOptions({ data: [] });
+        }
+    }
+
+    private refreshRows(): void {
+        const rows = this.table.getRowModel().rows;
+        this.tableData = rows.map(row => {
+            const rowState: TableRowState<TData> = { record: row.original };
+            return rowState;
+        });
     }
 
     private updateTableOptions(
@@ -75,6 +125,7 @@ export class Table<
     ): void {
         this.options = { ...this.options, ...updatedOptions };
         this.update(this.table.initialState);
+        this.refreshRows();
     }
 
     private readonly update = (state: TanStackTableState): void => {
@@ -91,8 +142,8 @@ export class Table<
         }));
     };
 
-    // Temporarily auto-detect the keys in TData to make columns.
-    // TODO: Remove this logic when another way to specify columns is provided.
+    // Generate columns for TanStack that correspond to all the keys in TData because all operations,
+    // such as grouping and sorting, will be performed on the data's records, not the values rendered within a cell.
     private generateColumns(): void {
         if (this.data.length === 0) {
             return;
@@ -110,8 +161,6 @@ export class Table<
         });
 
         this.updateTableOptions({ columns: generatedColumns });
-        this.columnHeaders = generatedColumns.map(x => x.header as string);
-        this.columns = this.columnHeaders;
     }
 }